refactored strength reducer tests
